+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [FIXED] Added special handling for requesting the `_users` database, as it can send `404` errors when a session expires.
+
 # 2.8.1 (2018-02-16)
 
 - [FIXED] Installation failures of 2.8.0 caused by missing VERSION file in distribution.